+++ afpfs-ng/lib/proto_attr.c
@@ -166,7 +166,7 @@ int afp_getextattr(struct afp_volume * volume, unsigned int dirid,
 	copy_path(server,p,pathname,strlen(pathname));
 	unixpath_to_afppath(server,p);
 	p2=p+sizeof_path_header(server)+strlen(pathname);
-	if (((unsigned int ) p2) & 0x1) p2++;
+	if (((unsigned long) p2) & 0x1) p2++;
 	req2=(void *) p2;
 
 	req2->len=htons(namelen);
